disdl/lambda/vision/create_vision_batch/requirements.txt
requests
pillow
redis
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
numpy<2.0
# zstandard
lz4
# torch==1.9.0+cpu
# torchvision==0.10.0+cpu